set up authentication and login session for user route handlers

+++ backend/controllers/campaigns.js
@@ -21,16 +21,10 @@ campaignsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body;
   const user = request.user;
 
-  const userDetails = {
-    id: user._id.toString(),
-    name: user.name,
-    email: user.email
-  }
-
   const campaign = new Campaign ({
     title: body.title,
     description: body.description,
-    userCreated: userDetails,
+    userCreated: user._id.toString(),
     amountTarget: body.amountTarget,
     dateStart: new Date().toISOString(),
     // dateEnd
@@ -50,8 +44,9 @@ campaignsRouter.put('/edit/:id', userExtractor, async (request, response) => {
   if (!campaign) {
     response.status(404).end()
   }
+  const campaignId = campaign.userCreated.toString();
 
-  if (campaign.userCreated.id === userId) {
+  if (campaignId === userId) {
     const updatedCampaign = await Campaign
       .findByIdAndUpdate(request.params.id, body, {new: true});
     if (updatedCampaign) {
@@ -63,9 +58,19 @@ campaignsRouter.put('/edit/:id', userExtractor, async (request, response) => {
 })
 
 campaignsRouter.delete('/:id', userExtractor, async (request, response) => {
-  // define user in current session
-  const deletedCampaign = await Campaign.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const userId = user._id.toString();
+  const campaign = await Campaign.findById(request.params.id)
+  if (!campaign) {
+    response.status(404).end()
+  }
+
+  if (campaign.userCreated.toString() === userId) {
+    await Campaign.findByIdAndRemove(request.params.id)
+    const newCampaignsCreated = user.campaignsCreated.filter(id => id.toString() !== campaign._id.toString())
+    await User.findByIdAndUpdate(userId, { campaignsCreated: newCampaignsCreated }, { new: true })
+    response.status(204).end()
+  }
 })
 
 module.exports = campaignsRouter;
